+++ DAL/CeilingService/LKA258Service.cs
@@ -71,7 +71,7 @@ namespace DAL
                     LKA258Id = Convert.ToInt32(objReader["LKA258Id"]),
                     ModuleTreeId = Convert.ToInt32(objReader["ModuleTreeId"]),
                     //最好不要用=null去判断，提示类型转换错误
-                    Length = objReader["Length"].ToString().Length == 0 ? 0 : Convert.ToDecimal(objReader["Length"]),
+                    Length = objReader["Length"].ToString().Length == 0 ? 0 : Convert.ToDouble(objReader["Length"]),
                     LightType = objReader["LightType"].ToString().Length == 0 ? "" : objReader["LightType"].ToString(),
                     Japan = objReader["Japan"].ToString().Length == 0 ? "" : objReader["Japan"].ToString(),
                 };
