+++ DAL/CeilingService/LKSSPECService.cs
@@ -74,8 +74,8 @@ namespace DAL
                     LKSSPECId = Convert.ToInt32(objReader["LKSSPECId"]),
                     ModuleTreeId = Convert.ToInt32(objReader["ModuleTreeId"]),
                     //最好不要用=null去判断，提示类型转换错误
-                    Length = objReader["Length"].ToString().Length == 0 ? 0 : Convert.ToDecimal(objReader["Length"]),
-                    Height = objReader["Height"].ToString().Length == 0 ? 0 : Convert.ToDecimal(objReader["Height"]),
+                    Length = objReader["Length"].ToString().Length == 0 ? 0 : Convert.ToDouble(objReader["Length"]),
+                    Height = objReader["Height"].ToString().Length == 0 ? 0 : Convert.ToDouble(objReader["Height"]),
                     WBeam = objReader["WBeam"].ToString().Length == 0 ? "" : objReader["WBeam"].ToString(),
                     SidePanel = objReader["SidePanel"].ToString().Length == 0 ? "" : objReader["SidePanel"].ToString(),
                     LightType = objReader["LightType"].ToString().Length == 0 ? "" : objReader["LightType"].ToString(),
